+++ lib/eb_deployer/application.rb
@@ -14,7 +14,7 @@ module EbDeployer
     def create_version(version_label, package)
       create_application_if_not_exists
 
-      source_bundle = if File.exists?(package)
+      source_bundle = if File.exist?(package)
                         if File.extname(package) == '.yml'
                           YAML.load(File.read(package))
                         else
@@ -26,7 +26,7 @@ module EbDeployer
                         bucket, obj_key = package.split(':')
                         {'s3_bucket' => bucket, 's3_key' => obj_key}
                       else
-                        raise "Unknown package file/format: #{package.inspect}"
+                        raise "Neither the file exists nor it is a valid s3 url: #{package.inspect}"
                       end
 
       unless @eb_driver.application_version_labels(@name).include?(version_label)