+++ lib/eb_deployer/event_poller.rb
@@ -45,7 +45,7 @@ module EbDeployer
     def digest(event)
       return nil unless event
       event = event.to_h if event.respond_to?(:to_h)
-      JSON.dump(event)
+      event.to_json
     end
   end
 end