+++ test/blue_green_deploy_test.rb
@@ -34,9 +34,9 @@ class BlueGreenDeployTest < DeployTest
       do_deploy(version_label, :smoke_test => smoke_test)
     end
 
-    assert_equal ['simple-production.elasticbeanstalk.com',
-                  'simple-production-inactive.elasticbeanstalk.com',
-                  'simple-production-inactive.elasticbeanstalk.com'], smoked_host
+    assert_equal ['simple-production.us-west-1.elasticbeanstalk.com',
+                  'simple-production-inactive.us-west-1.elasticbeanstalk.com',
+                  'simple-production-inactive.us-west-1.elasticbeanstalk.com'], smoked_host
   end
 
   def test_blue_green_deploy_should_blue_green_terminate_inactive_env_if_blue_green_terminate_inactive_is_enabled